html/cross-origin-opener-policy/reporting/access-reporting/property-fallbacks.https.html

<!doctype html>
<meta charset=utf-8>
<title> Check reports are not sent for "authorProp", "then", @@toStringTag, @@hasInstance, @@isConcatSpreadable </title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/get-host-info.sub.js></script>
<script src="/common/utils.js"></script>
<script src="/html/cross-origin-opener-policy/resources/dispatcher.js"></script>
<script>

const directory = "/html/cross-origin-opener-policy";
const executor_path = directory + "/resources/executor.html?pipe=";
const coep_header = '|header(Cross-Origin-Embedder-Policy,require-corp)';

let origin = [
  ["cross-origin" , get_host_info().HTTPS_REMOTE_ORIGIN ] ,
  ["same-site"    , get_host_info().HTTPS_ORIGIN        ] ,
];

let testCase = [
//operation                              ,
  w => w.authorProp                      , // throws a SecurityError
  w => w.then                            , // doesn't throw
  w => Object.prototype.toString.call(w) , // @@toStringTag, doesn't throw
  w => w instanceof Object               , // @@hasInstance, doesn't throw
  w => [].concat(w)                      , // @@isConcatSpreadable, doesn't throw
];

origin.forEach(([origin_name, origin]) => {
  testCase.forEach((op) =>  {
    promise_test(async t => {
      const opener_token = token();
      const openee_token = token();

      const openee_url = origin+ executor_path + `&uuid=${openee_token}`;
      const openee = window.open(openee_url);
      t.add_cleanup(() => send(openee_token, "window.close()"))

      // 1. Create an iframe in the openee.
      send(openee_token, `
        window.authorProp = "something";
        window.then = "something";
        let iframe = document.createElement("iframe");
        document.body.appendChild(iframe);

        send("${opener_token}", "openee loaded");
      `);
      let reply = await receive(opener_token);
      assert_equals(reply, "openee loaded");

      // 2. Try to access the openee.
      let observer = new ReportingObserver(()=>{});
      observer.observe();
      try {op(openee)} catch(e) {}
      let reports = observer.takeRecords();
      observer.disconnect();

      // 3. Check the received reports.
      assert_equals(reports.length, 0);

    }, `${origin_name} > ${op}`);
  });
});

</script>
